+++ src/main/java/com/estudiantes/repository/EjercicioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EjercicioRepository extends JpaRepository<Ejercicio, Long> {
+public interface EjercicioRepository extends JpaRepository<Ejercicio, Integer> {
 }
