+++ src/main/java/com/estudiantes/repository/OpcionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OpcionRepository extends JpaRepository<Opcion, Long> {
+public interface OpcionRepository extends JpaRepository<Opcion, Integer> {
 }
