+++ src/main/java/com/estudiantes/repository/PracticaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PracticaRepository extends JpaRepository<Practica, Long> {
+public interface PracticaRepository extends JpaRepository<Practica, Integer> {
 }
